refactor for extensibility/frameworkiness
