+++ ssh_server/handlers/handlers.go
@@ -1,0 +1,6 @@
+package handlers
+
+import "github.com/tbotnz/cisgo-ios/fakedevices"
+
+// PlatformHandler defines a default type for all platform handlers
+type PlatformHandler func(*fakedevices.FakeDevice)
